% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_dow}
\alias{add_dow}
\title{Get day of week.}
\usage{
add_dow(df)
}
\arguments{
\item{df}{Data frame with column `date`.}
}
\value{
Data frame with additional column `dow` representing the day of the week.
}
\description{
Get day of week.
}
