% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_is_christmas}
\alias{add_is_christmas}
\title{Check whether date is around Christmas.}
\usage{
add_is_christmas(df)
}
\arguments{
\item{df}{Data frame with column `date`.}
}
\value{
Data frame with additional column `is_christmas` representing whether the
date is on the 24th, 25th or 26th December.
}
\description{
Check whether date is around Christmas.
}
