% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_is_fom}
\alias{add_is_fom}
\title{Check whether date is the first of the month.}
\usage{
add_is_fom(df)
}
\arguments{
\item{df}{Data frame with column `date`.}
}
\value{
Data frame with additional column `is_fom` representing whether the date
is the first of the month.
}
\description{
Check whether date is the first of the month.
}
