% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_is_nyd}
\alias{add_is_nyd}
\title{Check whether date is New Years Day.}
\usage{
add_is_nyd(df)
}
\arguments{
\item{Data}{frame with column `date`.}
}
\value{
Data frame with additional column `is_nyd` representing whether the date
is on 1st January.
}
\description{
Check whether date is New Years Day.
}
