% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_prev_day}
\alias{add_prev_day}
\title{Get previous day's crime count.}
\usage{
add_prev_day(df)
}
\arguments{
\item{df}{Data frame with columns `date` and `n`.}
}
\value{
Data frame with counts from previous day added as column `n_pre`.
}
\description{
Get previous day's crime count.
}
