% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification_sensitivity}
\alias{classification_sensitivity}
\title{Sensitivity}
\usage{
classification_sensitivity(y_hat, y_test)
}
\arguments{
\item{y_hat}{Binary vector of predictions.}

\item{y_test}{Binary vector of observations.}
}
\value{
float Sensitivity.
}
\description{
Calculates the sensitivity (proportion of positives correctly identified).
}
\examples{
classification_sensitivity(c(0, 1, 0), c(0, 1, 1))
}
