% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community_bounds}
\alias{community_bounds}
\title{Community area boundaries
City of Chicago community area boundaries.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 10 columns.
}
\usage{
community_bounds
}
\description{
Community area boundaries
City of Chicago community area boundaries.
}
\references{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Community-Areas-current-/cauq-8yn6}
}
\keyword{datasets}
