% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dates}
\alias{convert_dates}
\title{Convert dates to other formats in dataframe}
\usage{
convert_dates(df, as_factors = TRUE, exclude = NULL)
}
\arguments{
\item{df}{Data frame containing `date` column to extract instants from.}

\item{as_factors}{Whether to convert `month`, `week`, `day` and `hour`
into factors.}

\item{exclude}{Specifies which instants to exclude from data frame returned.}
}
\value{
Data frame with `date` converted into instants.
}
\description{
Convert dates to other formats in dataframe
}
