% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_eval}
\alias{cv_eval}
\title{Indexed cross-validation for R6 class}
\usage{
cv_eval(object, X, y, error_funcs, index, ...)
}
\arguments{
\item{object}{R6 class with fit and predict methods.}

\item{X}{Data matrix}

\item{y}{Vector of observations.}

\item{error_funcs}{List of functions to compute test errors.}

\item{index}{Vector of indexes corresponding to test data.}

\item{...}{Additional arguments passed to fit}
}
\value{
Test error.
}
\description{
Computes cross-validation error using specified indexes (rows) of the data set
as the test set.
}
