% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_R6}
\alias{cv_R6}
\title{Cross-validation error for R6 class
Standard cross validation without folds.}
\usage{
cv_R6(object, X, y, error, test_size)
}
\arguments{
\item{object}{R6 class with fit and predict methods}

\item{X}{Matrix of data}

\item{y}{Vector of observations}

\item{error}{Function to calculate error (taking y and y_hat)}

\item{test_size}{Proportion of data to use for testing}
}
\value{
Result of error function
}
\description{
Cross-validation error for R6 class
Standard cross validation without folds.
}
\examples{
n <- 50
X <- matrix(runif(n, -2,4), nrow=n, ncol=1)
y <- as.vector(sin(X)) + rnorm(n, sd = 0.3)
kr <- KernelRidge$new("rbf", lambda = 1, 3)
cv_R6(kr, X, y, squared_error_loss, 0.2)
}
