% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_R6_idxs}
\alias{cv_R6_idxs}
\title{Indexed cross-validation for R6 class
Computes cross validation using specific indexes (rows) of the data set
as the test set.}
\usage{
cv_R6_idxs(object, X, y, error, idxs)
}
\arguments{
\item{object}{R6 class with fit and predict methods.}

\item{X}{Matrix of data.}

\item{y}{Vector of observations.}

\item{error}{Function to assess error.}

\item{idxs}{Vector of indices corresponding with test data.}
}
\value{
Average error across the folds.
}
\description{
Indexed cross-validation for R6 class
Computes cross validation using specific indexes (rows) of the data set
as the test set.
}
