% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cv_R6_k_fold}
\alias{cv_R6_k_fold}
\title{K-fold cross validation for R6 class}
\usage{
cv_R6_k_fold(object, X, y, error, k)
}
\arguments{
\item{object}{R6 class with fit and predict methods}

\item{X}{Matrix of data}

\item{y}{Vector of observations}

\item{error}{Function to calculate error (taking y and y_hat)}

\item{k}{Number of folds to use}
}
\value{

}
\description{
K-fold cross validation for R6 class
}
\examples{
n <- 50
X <- matrix(runif(n, -2,4), nrow=n, ncol=1)
y <- as.vector(sin(X)) + rnorm(n, sd = 0.3)
kr <- KernelRidge$new("rbf", lambda = 1, 3)
cv_R6_k_fold(kr, X, y, squared_error_loss, 5)
}
