% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{district_bounds}
\alias{district_bounds}
\title{Police district boundaries
City of Chicago police district boundaries.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 25 rows and 3 columns.
}
\usage{
district_bounds
}
\description{
Police district boundaries
City of Chicago police district boundaries.
}
\references{
\url{https://data.cityofchicago.org/Public-Safety/Boundaries-Police-Districts-current-/fthy-xz3r}
}
\keyword{datasets}
