% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_count_data}
\alias{get_count_data}
\title{Aggregate given dataset into count data.}
\usage{
get_count_data(
  df,
  time_period = "week",
  region = "community_area",
  crime_type = NULL
)
}
\arguments{
\item{df}{Dataset with columns corresponding to `time_period`, `region` and
`crime_type` (if desired), as well as an additional column `year`.}

\item{time_period}{The time period over which to aggregate.}

\item{region}{The region over which to aggregate.}

\item{crime_type}{The crime type over which to aggregate, if desired.}
}
\value{
Data frame `count_data` containing count data aggregated over the desired groups.
}
\description{
Aggregate given dataset into count data.
}
