% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kfold_cv}
\alias{kfold_cv}
\title{K-fold cross-validation for R6 class with parallel computation}
\usage{
kfold_cv(
  object,
  X,
  y,
  error_funcs,
  k,
  n_reps = 1000,
  parallel = FALSE,
  n_threads = NULL
)
}
\arguments{
\item{object}{R6 object with fit and predict methods.}

\item{X}{Data matrix.}

\item{y}{Response vector.}

\item{error_funcs}{List of functions for calculating error.}

\item{n_reps}{Number of repeats.}

\item{parallel}{Whether to compute in parallel.}

\item{n_threads}{The number of parallel threads to use. If NULL, this is
chosen to be the number of cores minus one.}
}
\value{
List of length equal to that of `error_funcs` with each element
containing a vector of length `n_reps` corresponding to the mean error
averaged over `k` folds.
}
\description{
K-fold cross-validation for R6 class with parallel computation
}
