% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Download Chicago Crime data using Socrata API}
\usage{
load_data(year = NULL, strings_as_factors = TRUE, drop_location = TRUE)
}
\arguments{
\item{year}{integer in 2001:2020 specifying the desired year.}

\item{strings_as_factors}{Whether to convert iucr, primary_type,
description and location_description to factors.}
}
\value{
tibble Data frame of Chicago Crime data.
}
\description{
Returns a data frame of Chicago Crime data from specified year, or entire dataset
from 2001 to present if no year is specified.
}
