% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic Regression.}
\description{
R6 class for Logistic Regression
}
\examples{
X <- subset(mtcars, select = c("mpg", "wt"))
y <- mtcars$am
lr <- LogisticRegression$new()
lr$fit(X, y)
lr$predict(X)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{Regularisation parameter (default = 0).}

\item{\code{solver}}{Solver to use in `optim`: one of "BFGS", "CG" or "L-BFGS-B".}

\item{\code{X_train}}{Training data matrix.}

\item{\code{y_train}}{Training response vector.}

\item{\code{X_test}}{Test data matrix.}

\item{\code{theta}}{The fitted parameters of the logistic regression model.}

\item{\code{control}}{List of control parameters to be passed to `optim`` control.}

\item{\code{rounding}}{Logical, whether to round predicted values to 0 and 1.}

\item{\code{predictions}}{The predicted values for `X_test`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogisticRegression$new()}}
\item \href{#method-fit}{\code{LogisticRegression$fit()}}
\item \href{#method-predict}{\code{LogisticRegression$predict()}}
\item \href{#method-clone}{\code{LogisticRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new `LogisticRegression` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$new(
  solver = "L-BFGS-B",
  lambda = 0,
  control = NULL,
  rounding = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solver}}{`"L-BFGS-B"`, `"BFGS"` or `"CG"`. Default is `"L-BFGS-B"`.}

\item{\code{lambda}}{Regularisation parameter, defualt 0.}

\item{\code{control}}{List of control parameters to be passed to `optim` control.}

\item{\code{rounding}}{Whether to round predicted values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the object to training data X and y.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$fit(X, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Training data (dataframe or matrix).}

\item{\code{y}}{Training data (vector).}

\item{\code{...}}{Additional arguments passed to optim.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict on X.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$predict(X, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{X training or testing X.}

\item{\code{quiet}}{If TRUE, the function will not return the predicted values and
will only update the object.
Computes log loss (with regularisation). Assumes one-hot-encoded
categorical variables and bias term in X.
Computes the gradient of the log loss function (with regularisation).
Assumes one-hot-encoded categorical variables and bias term in X.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
