% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_X}
\alias{parse_X}
\title{Parse matrix into useful format for ML algorithms}
\usage{
parse_X(X)
}
\arguments{
\item{X}{matrix or data.frame}
}
\value{
matrix
}
\description{
If a matrix is passed, function checks that it is not a character matrix.
If a data frame is passed, the function similarly checks the types, and
one hot encodes factors.
}
