% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{PoissonGAM}
\alias{PoissonGAM}
\title{Poisson GAM Regression}
\description{
R6 class wrapper for the Poisson family of GAMs, fitted via `gam` from the package
**mgcv** and tailored for use with a subset of the Chicago Crime dataset. To initialise
the object, the user supplies the variables over which to aggregate the data and, if 
"community_area" is chosen as `region`, whether to include a neighbourhood list as part
of the model. Given a matrix with columns corresponding to `time_period`, `region`, 
and `crime_type` (the latter two being optional), and a response vector containing the
counts of reported crimes aggregated over these variables, it will fit a GAM with a 
cyclic cubic regression smoother for`time_period` and treats `region` and `crime_type` 
as factors (unless a neighbourhood lists included, in which case a Markov random field 
smoother is applied to `region`). For more details on model fitting methods, see **mgcv**.
}
\examples{
df <- convert_dates(load_data(2019, na_omit = TRUE), filter_week = TRUE)
count_df <- get_count_data(df)
test_index <- sample(1:nrow(count_df), size = nrow(count_df) \%/\% 10)
X_train <- count_df[-test_index, 1:3]
y_train <- count_df[-test_index, 4]
X_test <- count_df[test_index, 1:3]
pg <- PoissonGAM$new(time_period = "week", region = "community_area",
                     crime_type = NULL, include_nb = FALSE)
pg$fit(X_train, y_train, n_threads = 1)
pred <- pg$predict(X_test)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X_train}}{The training data matrix.}

\item{\code{y_train}}{The training response vector.}

\item{\code{X_test}}{The test data matrix.}

\item{\code{time_period}}{Specifies the time period: must be one of `"yday"`, `"week"` or 
`"month"`.}

\item{\code{region}}{Specifies the spatial region to use: must be one of `"beat"`, 
`"district"`, `"ward"` or `"community"`. If `NULL`,  this will not be included in the model.}

\item{\code{crime_type}}{The crime type to include in the model as predictor: must be one of
`"primary_type"`, `"description"`, or `"fbi_code"`. If `NULL`, this will not be included 
in the model.}

\item{\code{include_nb}}{If `"community_area"` is the specified region, whether to include a 
neighbourhood list and then use a Markov random field smoother for fitting.}

\item{\code{include_year}}{Whether to include `year` as a predictor in the model.}

\item{\code{nb_list}}{Neighbourhood list for the community areas.}

\item{\code{gam_fitted}}{The GAM, fitted using package **mgcv**.}

\item{\code{fit_summary}}{Summary of the fitted GAM.}

\item{\code{predictions}}{Vector of predicted response values for the test dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PoissonGAM$new()}}
\item \href{#method-fit}{\code{PoissonGAM$fit()}}
\item \href{#method-predict}{\code{PoissonGAM$predict()}}
\item \href{#method-clone}{\code{PoissonGAM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new PoissonGAM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$new(
  time_period = "week",
  region = NULL,
  crime_type = NULL,
  include_nb = FALSE,
  include_year = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_period}}{One of "week", "month" or "yday" specifying the time period
over which counts are aggregated.}

\item{\code{region}}{One of "beat" or "community_area", specifying the region over
which counts are aggregated.}

\item{\code{crime_type}}{crime_type The crime type to include in the model as predictor. 
If `NULL`, this will not be included in the model.}

\item{\code{include_nb}}{Whether to include a neighbourhood list for the regions,
and hence whether to use a Markov random field smoother for the region.}

\item{\code{include_year}}{Whether to include `year` as a predictor in the model.}

\item{\code{filter_week}}{Whether to filter out the 53rd week.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Function for fitting a GAM to the training dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$fit(X_train, y_train, n_threads = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_train}}{The training dataset.}

\item{\code{y_train}}{The training response vector.}

\item{\code{n_threads}}{The number of threads to use for parallel smoothing
parameter selection methods in `gam`.}

\item{\code{...}}{Additional arguments to be passed to `gam`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Prediction for a new dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$predict(X_test = NULL, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_test}}{The test data matrix.}

\item{\code{quiet}}{If TRUE, the function will not return the predicted values and
will only update the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Vector of predicted values.
Check parameters are defined properly
Construct a neighbourhood list for community areas
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
