% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{r_squared}
\alias{r_squared}
\title{R-Squared}
\usage{
r_squared(y_hat, y)
}
\arguments{
\item{y_hat}{Vector of predictions.}

\item{y}{Vector of observed values.}
}
\value{
Float R2 value between 0 and 1.
}
\description{
Calculates the R2 value.
}
\examples{
r_squared(c(1,2,3), c(2,3,2))
}
