% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{rmse_loss}
\alias{rmse_loss}
\title{Root mean squared error loss}
\usage{
rmse_loss(y_hat, y)
}
\arguments{
\item{y_hat}{Vector of predicted values.}

\item{y}{Vector of observed values.}
}
\value{
Float root mean squared error loss
}
\description{
Root mean squared error loss
}
\examples{
rmse_loss(c(1,2,3), c(2,3,2))
}
