% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{sigmoid}
\alias{sigmoid}
\title{Sigmoid function}
\usage{
sigmoid(z)
}
\arguments{
\item{z}{numeric Log ratio estimate.}
}
\value{
float
}
\description{
Computes value of sigmoid function at `z`.
}
\examples{
sigmoid(0)
}
