% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{squared_error_loss}
\alias{squared_error_loss}
\title{Mean squared error loss}
\usage{
squared_error_loss(y_hat, y)
}
\arguments{
\item{y_hat}{Vector of predictions}

\item{y}{Vector of observed values}
}
\value{
Float of the mean squared error loss
}
\description{
Mean squared error loss
}
\examples{
squared_error_loss(c(1, 2, 3), c(2, 3, 2))
}
