% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{SupportVectorMachine}
\alias{SupportVectorMachine}
\title{Support vector machine wrapper}
\description{
R6 class wrapper for support vector machine implementation from e1071 package.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{kernel}}{Kernel function to use (as in svm documentation)}

\item{\code{fitted_model}}{`svm` object representing the fitted model.}

\item{\code{predictions}}{Predicted values of given test dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SupportVectorMachine$new()}}
\item \href{#method-fit}{\code{SupportVectorMachine$fit()}}
\item \href{#method-predict}{\code{SupportVectorMachine$predict()}}
\item \href{#method-clone}{\code{SupportVectorMachine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new SVM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SupportVectorMachine$new(kernel = "radial")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{kernel}}{Kernel function to use.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the object to training data X and y.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SupportVectorMachine$fit(X, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Training data (dataframe or matrix).}

\item{\code{y}}{Training data (vector).}

\item{\code{...}}{Additional arguments to be passed to `svm`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict on X.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SupportVectorMachine$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{X training or testing X.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SupportVectorMachine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
