% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{test_data.rda
The community bounds.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50000 rows and 26 columns.
}
\usage{
test_data
}
\description{
test_data.rda
The community bounds.
}
\references{
\url{https://data.cityofchicago.org/Public-Safety/Crimes-2001-to-Present/ijzp-q8t2}
}
\keyword{datasets}
