% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ward_bounds}
\alias{ward_bounds}
\title{Ward boundaries
City of Chicago ward boundaries.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 50 rows and 4 columns.
}
\usage{
ward_bounds
}
\description{
Ward boundaries
City of Chicago ward boundaries.
}
\references{
\url{https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Wards-2015-/sp34-6z76}
}
\keyword{datasets}
