% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yday_float}
\alias{yday_float}
\title{yday float}
\usage{
yday_float(timestamp)
}
\arguments{
\item{timestamp}{Date in POSIXct}
}
\value{
Floating point numbers between 1 and 366.
}
\description{
Convert timestamp to floating point number between 1 and 366 representing day of year.
}
