% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beat_bounds}
\alias{beat_bounds}
\title{Beat bounds
The bounds for the police beats.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 277 rows and 5 columns.
}
\usage{
beat_bounds
}
\description{
Beat bounds
The bounds for the police beats.
}
\references{
\url{https://data.cityofchicago.org/Public-Safety/Boundaries-Police-Beats-current-/aerh-rz74}
}
\keyword{datasets}
