% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classification_accuracy}
\alias{classification_accuracy}
\title{Accuracy
Calculates the proportion of correctly classified predictions.}
\usage{
classification_accuracy(y_hat, y_test)
}
\arguments{
\item{y_hat}{0 1 vector of predictions.}

\item{y_test}{0 1 vector of observations.}
}
\value{
float proportion of correct predictions
}
\description{
Accuracy
Calculates the proportion of correctly classified predictions.
}
\examples{
classification_accuracy(c(0,1,0), c(0,1,1))
}
