% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{classification_specificity}
\alias{classification_specificity}
\title{Specificity
Calculates the specificity (proportion of negatives correctly identified).}
\usage{
classification_specificity(y_hat, y_test)
}
\arguments{
\item{y_hat}{0 1 vector of predictions.}

\item{y_test}{0 1 vector of observations.}
}
\value{
float specificity.
}
\description{
Specificity
Calculates the specificity (proportion of negatives correctly identified).
}
\examples{
classification_specificity(c(0,1,0), c(0,1,1))
}
