% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{community_bounds}
\alias{community_bounds}
\title{Community bounds
The community bounds.}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 77 rows and 10 columns.
}
\usage{
community_bounds
}
\description{
Community bounds
The community bounds.
}
\references{
\url{https://data.cityofchicago.org/Public-Safety/Crimes-2001-to-Present/ijzp-q8t2}
}
\keyword{datasets}
