% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_dates}
\alias{convert_dates}
\title{Convert dates to other formats in dataframe}
\usage{
convert_dates(df, date_col = "date", as_factors = TRUE, exclude = NULL)
}
\arguments{
\item{df}{Data frame containing a column with date-time entries to extract
instants from.}

\item{date_col}{The name of the column with date-time entries, defaults to "date".}

\item{as_factors}{Whether to convert `month`, `week`, `day` and `hour`
into factors.}

\item{exclude}{Specifies which instants to exclude from data frame returned.}
}
\value{
Data frame `df` with entries of `date_col` converted into instants.
}
\description{
Convert dates to other formats in dataframe
}
