% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_data}
\alias{load_data}
\title{Download Chicago Crime data using Socrata API}
\usage{
load_data(
  year = NULL,
  strings_as_factors = TRUE,
  drop_location = TRUE,
  na_omit = FALSE,
  select = NULL,
  limit = NULL,
  ...
)
}
\arguments{
\item{year}{integer in 2001:2021 specifying the desired year, or two integers
specifying endpoints to filter between.}

\item{strings_as_factors}{Whether to convert `iucr`, `primary_type`,
`description`, `location_description` and `fbi_code` to factors.}

\item{drop_location}{Whether to drop location as it duplicates data in latitude
and longitude.}

\item{na_omit}{Whether to omit observations with NA values.}

\item{limit}{Limit set on how many rows of data to obtain.}

\item{...}{Additional arguments to be passed to `read.socrata`.}
}
\value{
tibble data frame of Chicago Crime data.
}
\description{
Returns a data frame of Chicago Crime data from specified year or range of years,
or returns entire dataset from 2001 to present if no years are specified.
}
