% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{Logistic Regression.}
\description{
R6 class for Logistic Regression
}
\examples{
X <- subset(mtcars, select = c("mpg", "wt"))
y <- mtcars$am
lr <- LogisticRegression$new()
lr$fit(X, y)
lr$predict(X)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{Regularisation parameter (defualt=0).}

\item{\code{solver}}{"BFGS", "CG" or "L-BFGS-B".}

\item{\code{X}}{Training X (dataframe or matrix).}

\item{\code{y}}{Training y vector.}

\item{\code{theta}}{The fitted parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LogisticRegression$new()}}
\item \href{#method-log_loss}{\code{LogisticRegression$log_loss()}}
\item \href{#method-grad_log_loss}{\code{LogisticRegression$grad_log_loss()}}
\item \href{#method-fit}{\code{LogisticRegression$fit()}}
\item \href{#method-predict}{\code{LogisticRegression$predict()}}
\item \href{#method-print}{\code{LogisticRegression$print()}}
\item \href{#method-clone}{\code{LogisticRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new LogisticRegression object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$new(solver = "L-BFGS-B", lambda = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{solver}}{"L-BFGS-B", "BFGS" or "CG". Default "L-BFGS-B".}

\item{\code{lambda}}{regularisation parameter, defualt 0.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_loss"></a>}}
\if{latex}{\out{\hypertarget{method-log_loss}{}}}
\subsection{Method \code{log_loss()}}{
Computes log loss (with regularisation).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$log_loss(theta, X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{Coefficients.}

\item{\code{X}}{Dataframe or matrix.}

\item{\code{y}}{0,1 vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grad_log_loss"></a>}}
\if{latex}{\out{\hypertarget{method-grad_log_loss}{}}}
\subsection{Method \code{grad_log_loss()}}{
Computes the gradient of the log loss function (with regularisation).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$grad_log_loss(theta, X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{Coefficients.}

\item{\code{X}}{Dataframe or matrix.}

\item{\code{y}}{0,1 Vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the object to training data X and y.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$fit(X, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Training data (dataframe or matrix).}

\item{\code{y}}{Training data (vector).}

\item{\code{...}}{Additional arguments passed to optim.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict on X.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$predict(X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{X training or testing X.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print LogisticRegression object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LogisticRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
