% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{otherise}
\alias{otherise}
\title{Convert less common strings to "OTHER", to be used when `primary_type` or
`description` is used from the Chicago Crime dataset.}
\usage{
otherise(string_vec, n_threshold, print_summary = TRUE)
}
\arguments{
\item{string_vec}{Vector of strings.}

\item{n_threshold}{Threshold count below which will be converted to other.}

\item{print_summary}{Prints a summary of operation performed.}
}
\value{
Vector of strings with the least common strings as "OTHER".
}
\description{
Convert less common strings to "OTHER", to be used when `primary_type` or
`description` is used from the Chicago Crime dataset.
}
