% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{otherise}
\alias{otherise}
\title{Convert less common strings to other}
\usage{
otherise(string_vec, n_threshold, print_summary = TRUE)
}
\arguments{
\item{print_summary}{}
}
\value{
vector of strings
}
\description{
Convert less common strings to other
}
