% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{PoissonGAM}
\alias{PoissonGAM}
\title{Poisson GAM Regression}
\description{
R6 class for the Poisson family of GAMs.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df_train}}{The training dataset.}

\item{\code{df_test}}{The test dataset.}

\item{\code{time_period}}{Specifies the time period: must be one of "days", "weeks" or
"months".}

\item{\code{region}}{Specifies the spatial region to use: must be one of "beat", "district",
"ward" or "community".}

\item{\code{include_nb}}{Whether to include a neighbours list for the discrete spatial
regions and then use a Markov random field smoother for fitting.}

\item{\code{include_crimetype}}{Whether to include crime types as a predictor in the model.}

\item{\code{filter_week}}{If using weeks, whether to filter out week 53, as it
typically has fewer than 7 days and can lead to spurious results.}

\item{\code{n_threads}}{Number of threads to use when fitting the GAM in parallel.}

\item{\code{nbd_list}}{Neighbourhood list for the specified region.}

\item{\code{count_train}}{Count data for the training dataset, grouped by the specified
time period, region, and (optional) crime type.}

\item{\code{count_test}}{Count data for the test dataset, grouped by the specified time
period, region, and (optional) crime type.}

\item{\code{gam_fitted}}{The GAM, fitted using package mgcv.}

\item{\code{fit_summary}}{Summary of the fitted GAM.}

\item{\code{predictions}}{Vector of predicted response values for the test dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PoissonGAM$new()}}
\item \href{#method-fit}{\code{PoissonGAM$fit()}}
\item \href{#method-predict}{\code{PoissonGAM$predict()}}
\item \href{#method-clone}{\code{PoissonGAM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new PoissonGAM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$new(
  time_period = "week",
  region = "community_area",
  include_nb = FALSE,
  include_crimetype = FALSE,
  filter_week = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_period}}{One of "week", "month" or "yday" specifying the time period
over which counts are aggregated.}

\item{\code{region}}{One of "beat" or "community_area", specifying the region over
which counts are aggregated.}

\item{\code{include_nb}}{Whether to include a neighbourhood list for the regions,
and hence whether to use a Markov random field smoother for the region.}

\item{\code{include_crimetype}}{Whether to include crime type as a predictor.}

\item{\code{filter_week}}{Whether to filter out the 53rd week.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Function for fitting a GAM to the training dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$fit(df_train, convert = FALSE, n_threads = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_train}}{The training dataset.}

\item{\code{convert}}{Whether the column `date` should be converted to instants
or they are already present.}

\item{\code{n_threads}}{The number of threads to use for parallel smoothing
parameter selection methods in `gam`.}

\item{\code{...}}{Additional arguments to be passed to `gam`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Prediction for a new dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$predict(df_test = NULL, convert = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_test}}{The training dataset.}

\item{\code{convert}}{Whether the column `date` should be converted to instants
or they are already present.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
