% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{PoissonGAM}
\alias{PoissonGAM}
\title{Poisson GAM Regression}
\description{
R6 class for the Poisson family of GAMs.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{df_train}}{The training dataset.}

\item{\code{df_test}}{The test dataset.}

\item{\code{time_period}}{Specifies the time period: must be one of "days", "weeks" or
"months".}

\item{\code{region}}{Specifies the spatial region to use: must be one of "beat", "district",
"ward" or "community".}

\item{\code{include_nb}}{Whether to include a neighbours list for the discrete spatial 
regions and then use a Markov random field smoother for fitting.}

\item{\code{include_crimetype}}{Whether to include crime types as a predictor in the model.}

\item{\code{n_threads}}{Number of threads to use when fitting the GAM in parallel.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PoissonGAM$new()}}
\item \href{#method-fit}{\code{PoissonGAM$fit()}}
\item \href{#method-predict}{\code{PoissonGAM$predict()}}
\item \href{#method-clone}{\code{PoissonGAM$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new PoissonGAM object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$new(
  time_period = "week",
  region = "community_area",
  include_nb = FALSE,
  include_crimetype = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Function for fitting a GAM to the training dataset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$fit(df_train, n_threads = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df_train}}{The training dataset.}

\item{\code{n_threads}}{The number of threads to use for parallel smoothing
parameter selection methods in `gam`}

\item{\code{...}}{Additional arguments to be passed to `gam`}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Function for predicting new dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$predict(df_test = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonGAM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
