% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{RandomForest}
\alias{RandomForest}
\title{Random forest
R6 class wrapper for random forest. Uses randomForest package.}
\description{
Random forest
R6 class wrapper for random forest. Uses randomForest package.

Random forest
R6 class wrapper for random forest. Uses randomForest package.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X_train}}{Training data matrix.}

\item{\code{y_train}}{Training response vector.}

\item{\code{X_test}}{Test data matrix.}

\item{\code{fitted_model}}{Fitted randomForest object.}

\item{\code{predictions}}{Predicted response values for test dataset.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RandomForest$new()}}
\item \href{#method-fit}{\code{RandomForest$fit()}}
\item \href{#method-predict}{\code{RandomForest$predict()}}
\item \href{#method-clone}{\code{RandomForest$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create new RandomForest object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomForest$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Fit the object to training data X and y.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomForest$fit(X, y, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Training data matrix.}

\item{\code{y}}{Training response vector.}

\item{\code{...}}{Additional arguments to be passed to `randomForest`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
Predict on X.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomForest$predict(X, logical = TRUE, quiet = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Test data matrix.}

\item{\code{logical}}{Whether to return the predictions as logical values.}

\item{\code{quiet}}{If TRUE, the function will not return the predicted values and
will only update the object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector Vector of predicted values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RandomForest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
