% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{yday_float}
\alias{yday_float}
\title{yday float}
\usage{
yday_float(timestamp)
}
\arguments{
\item{timestamp}{Date in POSIXct}
}
\value{
Floating point numbers
}
\description{
Convert timestamp to float in 1-366
}
